Refactor DrawerMenu and -Button to subfolders of TopBar
