+++ src/components/TopBar/DrawerMenu/DrawerButton/index.js
@@ -1,0 +1,3 @@
+import DrawerButton from './DrawerButton';
+
+export default DrawerButton;
