+++ src/components/TopBar/DrawerMenu/index.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 import { injectIntl } from 'react-intl';
 import { withStyles } from '@material-ui/core';
 import DrawerMenu from './DrawerMenu';
-import { findUserLocation } from '../../redux/actions/user';
+import { findUserLocation } from '../../../redux/actions/user';
 import styles from './styles';
 
 const mapStateToProps = (state) => {
